requirements.txt
openai==2.14.0
pip==25.3
python-dotenv==1.2.1
streamlit==1.52.2